misc: replace the sync client with async client
